+++ src/options/App.vue
@@ -4,12 +4,14 @@
     </div>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import Vue from 'vue';
+
+export default Vue.extend({
     data() {
         return {};
     },
-};
+});
 </script>
 
 <style lang="scss" scoped>