+++ src/popup/popup.ts
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import App from './App';
+import App from './App.vue';
 
 /* eslint-disable no-new */
 new Vue({