+++ src/ui/extension-icon.test.ts
@@ -1,4 +1,4 @@
-import {ExtensionIcon, STATE_ACTIVE, STATE_DEFAULT, STATE_INACTIVE} from './extension-icon';
+import { ExtensionIcon, STATE_ACTIVE, STATE_DEFAULT, STATE_INACTIVE } from './extension-icon';
 
 describe('Extension icon', () => {
     it('constants values is correct', () => {