+++ CHANGELOG.md
@@ -9,8 +9,11 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 ### Changed
 
 - Done migrate to typescript [#116]
+- New settings storage implementation [#119]
+- Indent size now equals `2`
 
 [#116]:https://github.com/tarampampam/random-user-agent/issues/116
+[#116]:https://github.com/tarampampam/random-user-agent/issues/119
 
 ## v2.2.11